Refactored app.py into modular components and reorganized project structure
